requirements.txt
opencv-python
pandas
azure-cognitiveservices-vision-customvision
msrest
torch
torchvision
detectron2
